requirements-spark.txt
pyspark==3.5.1
delta-spark==3.2.0
pyarrow==15.0.2
boto3==1.34.162
kafka-python==2.0.2
